Make Cassandra\Rows implement \Countable
This fixes the Unit-Tests which rely on old PHP5 behaviour of
`count($object)` always being `1` if it does not implement `\Countable`.

The common approach would be a fix of the tests, but in this case, it
makes sense to add the interface to Cassandra\Rows. There already exists
a working `count()` method.