Add additional dev tooling via another Dockerfile and a Makefile
The Makefile defines targets to build and interact with 2 type of
images:

1) dev (build-dev-image, shell-dev)

This is just a build of the standard Dockerfile which contains all the
dependencies to deevelop ext-cassandra. The image does neither contain the code
of the extension or build it. The code is instead mounted at `/code`.

2) full (build-full-image, shell-full, test*)

This is based on the dev image, but also adds the code to the image
itself, builds the extension, installs it and installs all the test
dependencies.

Test can be run in the image via the shell, or one of the test targets.
Integration tests will be slow (especially the first one), as
bootstrapping the test CS instance is very slow.